update30/03

+++ src/main/java/trasua/Service/User/giohangServiceImpl.java
@@ -19,7 +19,7 @@ public class giohangServiceImpl implements IgiohangService {
 	}
 
 	@Override
-	public HashMap<Long, giohang> editCart(long id, int quanty, HashMap<Long, giohang> gh) {
+	public HashMap<Long, giohang> editCart(long id, long quanty, HashMap<Long, giohang> gh) {
 		return GiohangDao.editCart(id, quanty, gh);
 	}
